+++ packages/widgetbook_generator/lib/generators/app_generator.dart
@@ -74,6 +74,7 @@ class AppGenerator extends GeneratorForAnnotation<App> {
 
     final instance = ListInstance(
       instances: directories,
+      type: 'WidgetbookNode',
     ).toCode();
 
     return 'final directories = $instance;';